change numpy .astype(float32) to .astype(float)

+++ merlin/analysis/segment.py
@@ -348,7 +348,7 @@ class CellPoseSegment(FeatureSavingAnalysisTask):
         '''Save a stack of images as a tiff file.'''
         with self.dataSet.writer_for_analysis_images(self, filename_prefix, fov) as outputTif:
              for i in range(image_stack.shape[0]):
-                    outputTif.save(image_stack[i].astype(float32),
+                    outputTif.save(image_stack[i].astype(float),
                                    photometric='MINISBLACK',
                                    contiguous=True)
 
@@ -624,7 +624,7 @@ class CellPoseSegmentSingleChannel(FeatureSavingAnalysisTask):
         '''Save a stack of images as a tiff file.'''
         with self.dataSet.writer_for_analysis_images(self, filename_prefix, fov) as outputTif:
              for i in range(image_stack.shape[0]):
-                    outputTif.save(image_stack[i].astype(float32),
+                    outputTif.save(image_stack[i].astype(float),
                                    photometric='MINISBLACK',
                                    contiguous=True)
